Conflicts:
	Exodus/Exodus/Datas/config.txt

+++ Exodus/Exodus/PlayGame/Items/Unit.cs
@@ -60,7 +60,7 @@ namespace Exodus.PlayGame
                 Item item = null;
                 int currentHeuristic = 0,
                     tempHeuristic;
-                int limit = this.SightRange * 10;
+                int limit = (int)(this.SightRange * 10);
                 for (int i = 0; i < Map.ListItems.Count; i++)
                     if (Map.ListItems[i].IdPlayer != this.IdPlayer && Map.ListItems[i].pos != null && this.pos != null)
                     {
